Break out test_runner into a module
Now that I'm interested in expanding our test cases to include a variety
of testing paths this is clearly the first step to maintain any level
of sanity.
